+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Enhancement: Improvements to saving changes in settings menu
 - Enhancement: Add syntax highlighting to the file viewer
 - Enhancement: Only re-render the G-Code viewer scene if something has changed
+- Enhancement: Add Facing wizard
 - Enhancement: Support gamepads as pendants
 - Enhancement: Added iPhone support
 - Enhancement: Improve Gcode viewer toolbar buttons layout